feat: добавить лэйблы, переписать компоненты
