+++ blocks/popup/__field/popup__field.css
@@ -1,0 +1,3 @@
+.popup__field {
+  position: relative;
+}
