+++ src/components/utils.js
@@ -1,0 +1,9 @@
+function openPopup(popup) {
+  popup.classList.add('popup_opened');
+}
+
+function closePopup(popup) {
+  popup.classList.remove('popup_opened');
+}
+
+export { openPopup, closePopup };
