+++ src/components/validate.js
@@ -1,0 +1,2 @@
+const form = document.querySelector('.edit-popup__edit-form');
+console.log(form);
