+++ src/pages/index.css
@@ -16,6 +16,8 @@
 @import url(../../blocks/popup/__photo/popup__photo.css);
 @import url(../../blocks/popup/__container/_type/popup__container_type-form.css);
 @import url(../../blocks/popup/__container/_type/popup__container_type-img.css);
+@import url(../../blocks/popup/__input-error/popup__input-error.css);
+@import url(../../blocks/popup/__field/popup__field.css);
 
 @import url(../../blocks/header/header.css);
 @import url(../../blocks/header/__logo/header__logo.css);
